Set qterminal window size to upstream default. (LP: #1786978)
Summary: 800x600 is more appropriate than 640x480, which was causing problems. This is part two of a fix for this bug. The previous one (D3) used the right terminal. This one fixes the display issue.

Test Plan:
 1. Right click on nm-tray
 2. Select "Edit Connections"

Assuming the other fix is present, should display nmtui-edit fully in a window.

Reviewers: tsimonq2

Differential Revision: https://phab.lubuntu.me/D6